Lesson_10/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pytest==8.2.0
requests==2.31.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
